+++ examples/browserify-esm/cypress.config.mjs
@@ -1,12 +1,16 @@
 import { defineConfig } from "cypress";
+import browserify from "@cypress/browserify-preprocessor";
 import { addCucumberPreprocessorPlugin } from "@badeball/cypress-cucumber-preprocessor";
-import { preprocessor } from "@badeball/cypress-cucumber-preprocessor/browserify";
+import { preprendTransformerToOptions } from "@badeball/cypress-cucumber-preprocessor/browserify";
 
 export async function setupNodeEvents(on, config) {
   // This is required for the preprocessor to be able to generate JSON reports after each run, and more,
   await addCucumberPreprocessorPlugin(on, config);
 
-  on("file:preprocessor", preprocessor(config));
+  on(
+    "file:preprocessor",
+    browserify(preprendTransformerToOptions(config, browserify.defaultOptions)),
+  );
 
   // Make sure to return the config object as it might have been modified by the plugin.
   return config;